% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_mapping.R
\name{run_spatial_mapping}
\alias{run_spatial_mapping}
\title{run_spatial_mapping}
\usage{
run_spatial_mapping(sp1, sp2, sp1_name, sp2_name)
}
\arguments{
\item{sp1}{SpatialPolygonsDataFrame}

\item{sp2}{SpatialPolygonsDataFrame}

\item{sp2_name}{Name of field in sp2 for mapping}
}
\description{
maps two spatial polygons by
calculating the proportion of area overlapping
}
\author{
Jennifer An
}
